metropolis 0.1.0 - initial release
A key-value store for Rack.  It emulates the basic REST
functionality (PUT/GET/DELETE) of TokyoTyrant as a Rack handler.
TDB is the primary DBM supported but TokyoCabinet (hash) is
supported as well.

For TDB, it is multi-process safe, but is currently not
thread-safe due to limitations of the upstream TDB library.